Flask App/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-magic==0.4.27
python-magic-bin==0.4.14
python-nmap==0.7.1
Werkzeug==3.0.1
